% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtmosphericEmissivity.R
\name{AtmosphericEmissivity}
\alias{AtmosphericEmissivity}
\title{Atmosphere Emissivity}
\usage{
AtmosphericEmissivity(airtemp, cloudiness, vp = NULL, opt = "linear")
}
\arguments{
\item{airtemp}{Air temperature}

\item{cloudiness}{Cloudness fraction}

\item{vp}{vapor pressure in kPa (optional)}

\item{opt}{Option to choose methodologies (Default='linear'). if opt='Brutsaert, vp is requried.}
}
\value{
Atmosphere Emissivity
}
\description{
Calculating atmosphere emissivity to be used for longwave estimation based on Campbell and Norman (1998)
}
\examples{
AtmosphericEmissivity(airtemp=-12.8, cloudiness=0.4)

}
\references{
Campbell, G.S., Norman, J.M., 1998. An Introduction to Environmental Biophysics, second ed., Springer, New York,p. 286.
Todd Walter, M., Brooks, E.S., McCool, D.K., King, L.G., Molnau, M., Boll, J., 2005. Process-based snowmelt modeling: does it require more input data than temperature-index modeling? Journal of Hydrology 300, 65–75. https://doi.org/10.1016/j.jhydrol.2004.05.002
}
