% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtmospherePressure.R
\name{Atmp.Pressure.Diurnal}
\alias{Atmp.Pressure.Diurnal}
\title{Atmosphere pressure}
\usage{
Atmp.Pressure.Diurnal(Elev, Tair, VPD = NULL)
}
\arguments{
\item{Elev}{Elevation in meter}

\item{Tair}{Temperature in degree Celcius}

\item{VPD}{Vapor pressure deficit (kPa); optional}
}
\value{
Atmp.pressure in hPa
}
\description{
Calculating atmosphere pressure in hPa
}
