% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dewpoint_Diurnal.R
\name{Dew.point.Diurnal}
\alias{Dew.point.Diurnal}
\title{Diurnal dew point temperature}
\usage{
Dew.point.Diurnal(Preday, Present, Afterday)
}
\arguments{
\item{Preday}{Sunrise time (hour) and dewpoint temperature (Celcius) on previous day, c(sunrise,dewpoint)}

\item{Present}{Sunrise time (hour) and dewpoint temperature (Celcius) on present day}

\item{Afterday}{Sunrise time (hour) and dewpoint temperature (Celcius) on after day}
}
\value{
Hourly dew point temperature in degree Celcius (24 points) for the present day
}
\description{
Calculate hourly dew point temperature given daily dew point temperature drived from daily Tmin and RH
It assumes that daily dew point occurs at sun rise time and dewpoint diurnal is calucated by linear interpolation as described in Bohn et al. (2013)
}
\references{
Bohn, T.J., Livneh, B., Oyler, J.W., Running, S.W., Nijssen, B., Lettenmaier, D.P., 2013. Global evaluation of MTCLIM and related algorithms for forcing of ecological and hydrological models. Agricultural and Forest Meteorology 176, 38–49. https://doi.org/10.1016/j.agrformet.2013.03.003
}
\author{
Hyung Eum, Government of Alberta
}
