% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dewpoint.R
\name{Dew.Point}
\alias{Dew.Point}
\title{Dewpoint temperature}
\usage{
Dew.Point(Tair, RH)
}
\arguments{
\item{Tair}{Air temperature in degree celcius}

\item{RH}{Relative humidity (ratio)}
}
\value{
Dew point temperature in degree Celcius
}
\description{
Calculating dewpoint temperature in degree Celcius given air temperature and relative humidity
}
\references{
Lawrence, Mark G., 2005: The relationship between relative humidity and the dewpoint temperature
in moist air: A simple conversion and applications. Bull. Amer. Meteor. Soc., 86, 225-233.
doi: http;//dx.doi.org/10.1175/BAMS-86-2-22
}
