% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Disaggregation.R
\name{Disaggregation}
\alias{Disaggregation}
\title{DB-HCOP}
\usage{
Disaggregation(
  lon,
  lat,
  Ele,
  tz = Sys.timezone(),
  Start.Date,
  End.Date,
  Prcp,
  Tmin,
  Tmax,
  Wind,
  RH,
  Solar.Rad = NULL,
  Long.Rad = NULL,
  Prcp.STN.List,
  Wind.STN.List,
  Dir.Prcp.ratio,
  Dir.Wind.ratio,
  N.Neighbor = 3,
  Upper.bound = 1.2,
  Lower.bound = 0.8
)
}
\arguments{
\item{lon}{Logitude in degree}

\item{lat}{Latitude in degree}

\item{Ele}{Elevatoin, a.s.m.l [m]}

\item{tz}{Time zone}

\item{Start.Date}{Start date in character (YYYY-MM-DD), ex:'2000-01-01'}

\item{End.Date}{End date in character (YYYY-MM-DD), ex:'2000-12-31'}

\item{Prcp}{Daily precipitation from Start.Date to End.Date [mm/day]}

\item{Tmin}{Daily minimum temperature from Start.Date to End.Date [degree Celcius]}

\item{Tmax}{Daily maximum temperature from Start.Date to End.Date [degree Celcius]}

\item{Wind}{Daily wind speed from Start.Date to End.Date [m/sec]}

\item{RH}{Daily relative humidity from Start.Date to End.Date [ratio]}

\item{Solar.Rad}{Daily solar(shortwave) radiation from Start.Date to End.Date [W/m2]}

\item{Long.Rad}{Daily longwave incoming radiation from Start.Date to End.Date [W/m2]}

\item{Prcp.STN.List}{List of precipitation stations providing hourly fraction of precipitation, (Station name,ID, Lat, Lon)}

\item{Wind.STN.List}{List of wind stations providing hourly fraction of wind speed, (Station name,ID, Lat, Lon)}

\item{Dir.Prcp.ratio}{Path of a folder that stores precipitation fraction files for all stations listed in Prcp.STN.List}

\item{Dir.Wind.ratio}{Path of a folder that stores wind fraction files for all stations listed in Wind.STN.List}

\item{N.Neighbor}{Number of nearest neighbor climate stations (default=3)}

\item{Upper.bound}{Upper bound for pupulating samples of houlry fractions from nearest neighbor climate stations (default=1.2)}

\item{Lower.bound}{Lower bound for pupulating samples of houlry fractions from nearest neighbor climate stations (default=0.8)}
}
\value{
Prcp.Diurnal[mm/hour],Air.Temp.Diurnal[degree Celcius],Dew.Temp.Diurnal[degree Celcius],
RH.Diurnal[ratio],SH.Diurnal[ratio],Wind.Diurnal[m/sec],Atmosphere.Diurnal[hPa],WVP.Diurnal[hPa],
Solar.Rad.Diurnal[W/m2],Long.Rad.Diurnal[W/m2],Cloudness.Diurnal[ratio]
}
\description{
Daily based-Hourly Climate Output Producer (DB-HCOP) disaggregating daily to hourly
}
\details{
Disaggregation of precipitation and wind speed employs hourly fraction sampled from observed hourly climate data.
A list of climate stations (i.e., Prcp.STN.List and Wind.STN.List) is required to find a set of nearest neighbor stations (N.Neighbor) from a target point (lon and lat).
With a fragment approach, a set of samples of hourly fraction is collected from the nearest neighbor stations, which has a similarity to daily precipitation and wind speed within upper and lower bounds.
Then a hourly fraction is randomly chosen from the populated samples of hourly fractions.
For diurnal temperature, radiation, dew point temperature, please refer to 'Temperature_Diurnal', 'Solar_Diurnal','Longwave_Diurnal','Dew.point.Diurnal'.
}
