% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cloudiness.R
\name{EstCloudiness}
\alias{EstCloudiness}
\title{Cloud fraction}
\usage{
EstCloudiness(
  Tx = (-999),
  Tn = (-999),
  trans = NULL,
  transMin = 0.15,
  transMax = 0.75,
  opt = "linear"
)
}
\arguments{
\item{Tx}{Daily maximum temperature}

\item{Tn}{Daily minimum temperature}

\item{trans}{Transmissivity. If trans=NULL, it can be estimated by 'transmissivity' function using daily min/max temperature}

\item{transMin}{Minimum transmissivity (default=0.15)}

\item{transMax}{Maximum transmissivity (default=0.75)}

\item{opt}{'linear' (default) or 'Black'}
}
\value{
Cloud fraction
}
\description{
Calculate daily cloud fraction given transmissivity
}
