% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fraction_Prcp_Wind.R
\name{Fraction}
\alias{Fraction}
\title{Hourly Fraction}
\usage{
Fraction(
  STN.name,
  STN.ID,
  Hourly.Data,
  Var.Name = "Prcp",
  Dir.Output,
  Output.format = "csv"
)
}
\arguments{
\item{STN.name}{Station name in character}

\item{STN.ID}{Station ID in character}

\item{Hourly.Data}{Hourly data (time and data), time format='YYYY-MM-DD hh:mm:ss'. Please refer to 'Hourly.Data.Sample.Prcp'}

\item{Var.Name}{Variable name ('Prcp' or 'Wind')}

\item{Dir.Output}{Output folder to store hourly fraction file generated}

\item{Output.format}{Output format: csv or xlsx (default='csv')}
}
\value{
Fraction.Var.calc.avail
}
\description{
Calculating hourly fraction of pecipitation and wind speed from hourly climate data
}
\details{
The hourly fractions are employed into 'Disaggregation' function to disaggre daily to hourly for precipitation and wind speed.
Hourly Fraction generates a fraction file named by 'STN.name'_'STN.ID'_Hourly_Fraction_'Var.Name'.csv (or xlsx).
}
\examples{
data(Hourly.Data.Sample.Wind)
STN.name="WILLOW CREEK 1"
STN.ID="306GE70"
Hourly.Data=Hourly.Data.Sample.Wind
Var.Name='Wind'
Dir.Output='Ouput'
Output.format='csv'
Fraction(STN.name,STN.ID,Hourly.Data,Var.Name,Dir.Output,Output.format)
}
