% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Longwave.R
\name{Longwave}
\alias{Longwave}
\title{Longwave}
\usage{
Longwave(emissivity, temp)
}
\arguments{
\item{emissivity}{Emissivity estimated by 'AtmosphericEmissivity' function}

\item{temp}{Air temperature in degree Celcius}
}
\value{
Daily longwave [W/m2]
}
\description{
Daily longwave radiation based on the Sephan-Boltzman equation [W/m2]
1 W/m2= 86.4 KJ/m2/day (1*3600*24*10^-3)
1 W/m2= 3.6 KJ/m2/hour (1*3600*10^-3)
1 W/m2= 0.001 KJ/m2/sec
1 W = 1 Joule/sec
}
\references{
Todd Walter, M., Brooks, E.S., McCool, D.K., King, L.G., Molnau, M., Boll, J., 2005. Process-based snowmelt modeling: does it require more input data than temperature-index modeling? Journal of Hydrology 300, 65–75. https://doi.org/10.1016/j.jhydrol.2004.05.002
}
