% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Longwave_Diurnal.R
\name{Longwave_Diurnal}
\alias{Longwave_Diurnal}
\title{Diurnal Longwave calculation}
\usage{
Longwave_Diurnal(Air.Temp.hour, cloudiness, Longwave.Day, emissivity = NULL)
}
\arguments{
\item{Air.Temp.hour}{Hourly air temperature (degree Celcius) from 'Temperature_Diurnal' function}

\item{cloudiness}{Cloud fraction from 'EstCloudiness' function}

\item{Longwave.Day}{Daily incoming longwave radiation [W/m2]}

\item{emissivity}{Emissivity derived from a Campbell and Norman method given air temperature (optional)}
}
\value{
Hourly longwave [W/m2]
}
\description{
Calculate hourly longwave radiation using the Stefan-Boltzmann equation given cloudiness estimated from 'Estcloundiness' with Tmin and Tmax
It assumes that the cloudiness is constant thourgh a day.
}
\details{
1 W/m2= 86.4 KJ/m2/day (1*3600*24*10^-3)
1 W/m2= 3.6 KJ/m2/hour (1*3600*10^-3)
1 W/m2= 0.001 KJ/m2/sec
1 W = 1 Joule/sec
}
