% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prcp_Wind_Diurnal.R
\name{Prcp_Wind_Diurnal}
\alias{Prcp_Wind_Diurnal}
\title{Disaggregation of daily to hourly}
\usage{
Prcp_Wind_Diurnal(
  lat,
  lon,
  Date,
  Var.Day,
  Upper.Ratio = 1.2,
  Lower.Ratio = 0.8,
  N.Neighbor = 3,
  List.ECCC.STN,
  Fraction.Folder,
  Fraction.format = "xlsx"
)
}
\arguments{
\item{lat}{Latitude in degree}

\item{lon}{Longitude in degree}

\item{Date}{Current date, format("YYYY-MM-DD")}

\item{Var.Day}{Daily precipitation(mm/day) or wind speed(m/s)}

\item{Upper.Ratio}{Upper bound ratio to select hourly fractions from nearest neighbor stations}

\item{Lower.Ratio}{Lower bound ratio to select hourly fractions from nearest neighbor stations}

\item{N.Neighbor}{Number of nearest neighbor climate stations (default=3)}

\item{List.ECCC.STN}{List of ECCC stations that hourly data is available}

\item{Fraction.Folder}{Location (Folder path) where hourly fraction files are located}

\item{Fraction.format}{File format in character: xlsx or csv (default=xlsx)}
}
\description{
Disaggregating daily to hourly for precipiation and wind using hourly fraction of precipitation drived from ECCC stations
}
\details{
Hourly fraction files are named by Station name and ID described in List.ECCC.STN.
If station name and ID are 'ABEE AGDM' and '3010010', for example, hourly fraction file name should be 'ABEE AGDM_3010010_Hourly_Fraction_Prcp.xlsx'
The format of List.ECCC.STN and hourly fraction is available in this package by data('List.ECCC.STN.Prcp')
}
\author{
Hyung Eum (hieum01@gov.ab.ca)
}
