% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Specific_Humidity.R
\name{SH}
\alias{SH}
\title{Specific humidity}
\usage{
SH(e, Tair, Ele)
}
\arguments{
\item{e}{partial water vapor pressure in hPa}

\item{Tair}{Temperature in degree Celcius}

\item{Elev}{Elevation in meter}
}
\value{
Specific humidity (SH) in ratio
}
\description{
Calculate specific humidity given partial water vapor pressure and air temperature
}
\examples{
Tair=25
e=WVP.RH.T(RH=0.7,Tair=Tair)
SH(e, Tair=Tair,Ele=1011)

}
\references{
Molecular weight of water vapor and dry air: https://en.wikipedia.org/wiki/Molar_mass
}
\seealso{
\code{\link{WVP.RH.T}}, \code{\link{WVP.Td}}
}
