% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solar_Diurnal.R
\name{Solar_Diurnal}
\alias{Solar_Diurnal}
\title{Diurnal solar radiation}
\usage{
Solar_Diurnal(lat, lon, date, tz = Sys.timezone(), Solar.Day)
}
\arguments{
\item{lat}{Latitude in degrees}

\item{lon}{Longitude in degrees}

\item{date}{Date in characer (ex: '2000-12-31')}

\item{tz}{Time zone}

\item{Solar.Day}{Daily solar radiation [W/m2]}
}
\value{
Hourly solar radiation [W/m2]
}
\description{
Diurnal Solar radiation with daily total solar radiation
}
\details{
Using a simple cosine bell (Hirschmann, 1974), dirunal solar radiation is calculated.
Given the dirunal solar radition, fractin of hourly solar radition is applied to daily solar ration to calculate hourly solar radiatio.
}
\examples{
Solar_Diurnal(lat=53.4,lon=-117.5,date='1961-01-01',tz='MST',Solar.Day=34.7)

}
\references{
Kimball, B.A., Bellamy, L.A., 1986. Generation of diurnal solar radiation, temperature, and humidity patterns. Energy in Agriculture 5, 185–197. https://doi.org/10.1016/0167-5826(86)90018-5
Hirschmann, J.R., 1974. The cosine function as a mathematical expression for the processes of solar energy. SOl. Energy, 16, 117-124.
}
