% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sunrise_Sunset_Hour.R
\name{Sun.rise.set.Hour}
\alias{Sun.rise.set.Hour}
\title{Sunrise and sunset time}
\usage{
Sun.rise.set.Hour(lat, lon, date, tz = Sys.timezone(), span = 1)
}
\arguments{
\item{lat}{Latitude in degree}

\item{lon}{Longitude in degree}

\item{date}{Date in character (YYYY-MM-DD), ex: '2023-04-30'}

\item{tz}{Time zone}

\item{span}{Squence days (default=1)}
}
\value{
Sun.Data: Date, Sunrise time, Sunset time, Lenght of day in hour
}
\description{
Calculating sunrise and sunset time (hour) corresponding to latitude and longitude based on sun-methods {maptools}
}
\examples{
Sun.rise.set.Hour(lat=53.4,lon=-117.5,date='1961-01-01',tz='MST',span=1)

}
\references{
Meeus, J. (1991) Astronomical Algorithms. Willmann-Bell, Inc.
Sunrise/Sunset: https://gml.noaa.gov/grad/solcalc/sunrise.html
}
