% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Temperature_Diurnal.R
\name{Temperature_Diurnal}
\alias{Temperature_Diurnal}
\title{Diurnal temperature}
\usage{
Temperature_Diurnal(
  lat,
  lon,
  date,
  tz = Sys.timezone(),
  Tmin,
  Tmax,
  preday = NULL,
  afterday = NULL
)
}
\arguments{
\item{lat}{Latitdue in degrees}

\item{lon}{Longitude in degrees}

\item{date}{Current date (YYYY-MM-DD) in character (ex: '2000-12-31')}

\item{tz}{Time zone}

\item{Tmin}{Daily minimum temperature in degree Celcius}

\item{Tmax}{Daily maximum temperature in degree Celcius}

\item{preday}{tmin/tmax, sunrise and sunset time (in minute) on the previous day: c(tmin, tmax, sunrise, sunset)}

\item{afterday}{tmin/tmax, sunrise and sunset time (in minute) on the after day: c(tmin, tmax, sunrise, sunset)}
}
\value{
Hourly temperature
}
\description{
Given daily min/max temperature, hourly temperature is calculated.
}
\details{
It assumes that minimum temperature occurs at sunrise time while maximum temperature occurs at T_rise+0.67(T_set-T_rise).
Once the times of min/max temperature on previous, currnet and after days are set, hourly temperature is calculated by a spline of 3rd Hermite polynomials.
}
\references{
Bohn, T.J., Livneh, B., Oyler, J.W., Running, S.W., Nijssen, B., Lettenmaier, D.P., 2013. Global evaluation of MTCLIM and related algorithms for forcing of ecological and hydrological models. Agricultural and Forest Meteorology 176, 38–49. https://doi.org/10.1016/j.agrformet.2013.03.003
}
