% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaporPressure.R
\name{WVP.Td}
\alias{WVP.Td}
\title{Partial water vapor pressure given dew point}
\usage{
WVP.Td(T.dew)
}
\arguments{
\item{T.dew}{Dew point in Kelvin (K) or in degree Celsius (°C)}
}
\value{
Partial vapor pressure in hectopascal (hPa)

artial water vapor pressure [hPa]
}
\description{
calculate partial water vapor pressure given dew point temperature
}
\seealso{
\code{\link{SVP.ClaCla}}.
}
